IGNITE-19670: Improve CatalogService test coverage.
