+++ modules/sql-engine/src/testFixtures/java/org/apache/ignite/internal/sql/engine/util/tpch/TpchTables.java
@@ -31,6 +31,7 @@ import java.util.Iterator;
 import org.apache.ignite.sql.ColumnType;
 import org.apache.ignite.sql.IgniteSql;
 import org.apache.ignite.table.KeyValueView;
+import org.apache.ignite.table.Tuple;
 
 /**
  * Enumeration of tables from TPC-H specification.
